% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charger_gpkg.R
\name{charger_gpkg}
\alias{charger_gpkg}
\title{charger_gpkg}
\usage{
charger_gpkg(layers = NULL, path = NULL)
}
\arguments{
\item{layers}{Liste de layer du geopackage}

\item{path}{chemin vers le fichier (choix de fichier si pas de chemin)}
}
\value{
Un ou plusieurs objets `sf` nommés comme les couches importées
}
\description{
Fonction permettant de charger rapidement les couches contenu dans un geopackage
}
\examples{
charger_gpkg(layers = c('Flore','Releve_Phyto'))
}
