% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copo.R
\name{copo}
\alias{copo}
\title{copo}
\usage{
copo()
}
\value{
Une connexion active à la base PostgreSQL (objet de type `DBIConnection`)
}
\description{
Fonction permettant de se connecter à la base de données PostgreSQL en utilisant un fichier `.Renviron`.
}
\details{
La fonction suppose que le fichier `.Renviron` contient les variables d’environnement `PGUSER` et `PGPASSWORD`, et que la base `BiodiversitySQL` est accessible.
}
\examples{
con = copo()
}
