% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigrammisation.R
\name{trigrammisation}
\alias{trigrammisation}
\title{trigrammisation}
\usage{
trigrammisation(
  nom_valide = taxref$nom_valide,
  lb_nom = taxref$lb_nom,
  rang = taxref$rang
)
}
\arguments{
\item{nom_valide}{Vecteur de caractères contenant les noms scientifiques valides (par défaut : `taxref$nom_valide`)}

\item{lb_nom}{Vecteur de caractères contenant le libellé du nom (par défaut : `taxref$lb_nom`)}

\item{rang}{Vecteur de caractères indiquant le rang taxonomique (par défaut : `taxref$rang`)}
}
\value{
Un vecteur de chaînes de caractères correspondant au trigramme attribué à chaque taxon
}
\description{
Génère un trigramme unique pour chaque taxon en se basant sur les règles de construction à partir du nom valide. Gère les cas de sous-espèces, variétés, formes, hybrides, et doublons. Permet également d'utiliser le nom complet pour les rangs supérieurs au genre.
}
\examples{
trigrammisation(nom_valide = c("Quercus ilex", "Quercus x pubescens"),
                lb_nom = c("Chêne vert", "Chêne hybride"),
                rang = c("ES", "ES"))
}
