% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_schema.R
\name{write_to_schema}
\alias{write_to_schema}
\title{write_to_schema}
\usage{
write_to_schema(con, schema, table, data, append = FALSE, overwrite = FALSE)
}
\arguments{
\item{con}{Connexion à la BDD (contruite via copo())}

\item{schema}{Schema de la BDD}

\item{table}{Schema de la BDD}

\item{data}{Schema de la BDD}

\item{append}{Pour ajouter des ligne si TRUE}

\item{overwrite}{Pour remplacer le tableau si TRUE}
}
\description{
Fonction permettant d'ajouter des lignes dans une BDD
}
\examples{
write_to_schema(con, "inaturalist", "corresp_taxref", new_corresp, append = TRUE)
}
